requirements.txt
langchain
langchain_ollama
selenium
beautifulsoup4
python-dotenv
discord
